+++ src/steps/resource-manager/policy/index.ts
@@ -8,15 +8,10 @@ import {
 import { createAzureWebLinker } from '../../../azure';
 import { IntegrationStepContext, IntegrationConfig } from '../../../types';
 import { ACCOUNT_ENTITY_TYPE, STEP_AD_ACCOUNT } from '../../active-directory';
-import { STEP_RM_SUBSCRIPTIONS } from '../subscriptions';
-import { STEP_RM_RESOURCES_RESOURCE_GROUPS } from '../resources';
-import {
-  PolicyEntities,
-  PolicyRelationships,
-  STEP_RM_POLICY_ASSIGNMENTS,
-} from './constants';
+import { PolicyEntities, PolicyRelationships, PolicySteps } from './constants';
 import { AzurePolicyClient } from './client';
 import { createPolicyAssignmentEntity } from './converters';
+import { getResourceManagerSteps } from '../../../getStepStartStates';
 
 export * from './constants';
 
@@ -45,6 +40,17 @@ export async function fetchPolicyAssignments(
             _class: RelationshipClass.HAS,
             from: scopeEntity,
             to: policyAssignmentEntity,
+            /**
+             * NOTE: Azure Management Groups, Azure Subscriptions, Azure Resource Groups, and Azure Resources can all have Azure Policy Assignments.
+             * We currently don't ingest Management Groups, but we do ingest Subscriptions, Resource Groups, and some Resources
+             * We've chosen to represent the relationship as ANY_SCOPE has a Policy Assignment.
+             * This is because we want the relationship metadata to line up with the actual ingested relationships.
+             * If we say that we expect ANY_SCOPE_had_policy_assignment and instead generate that a azure_storage_account_has_policy_assignment,
+             * it might cause issues down the road.
+             */
+            properties: {
+              _type: 'ANY_SCOPE_has_policy_assignment',
+            },
           }),
         );
       }
@@ -56,17 +62,13 @@ export const policySteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
   {
-    id: STEP_RM_POLICY_ASSIGNMENTS,
+    id: PolicySteps.POLICY_ASSIGNMENTS,
     name: 'Policy Assignments',
     entities: [PolicyEntities.POLICY_ASSIGNMENT],
-    relationships: [
-      PolicyRelationships.RESOURCE_GROUP_HAS_POLICY_ASSIGNMENT,
-      PolicyRelationships.SUBSCRIPTION_HAS_POLICY_ASSIGNMENT,
-    ],
+    relationships: [PolicyRelationships.ANY_RESOURCE_POLICY_ASSIGNMENT],
     dependsOn: [
       STEP_AD_ACCOUNT,
-      STEP_RM_SUBSCRIPTIONS,
-      STEP_RM_RESOURCES_RESOURCE_GROUPS,
+      ...getResourceManagerSteps().executeFirstSteps,
     ],
     executionHandler: fetchPolicyAssignments,
   },